src/RAG/requirements.txt
faiss==1.5.3
numpy==2.4.1
openai==2.15.0
pydantic==2.12.5
python-dotenv==1.2.1
sentence_transformers==5.2.0
torch==2.9.1
